+++ ft_strlen.c
@@ -6,13 +6,14 @@
 /*   By: rexposit <rexposit@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/17 18:04:27 by rexposit          #+#    #+#             */
-/*   Updated: 2024/09/19 16:27:15 by rexposit         ###   ########.fr       */
+/*   Updated: 2024/09/20 11:35:42 by rexposit         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stddef.h> //SUSTITUIR LIBRERIA CON LIBFT.H
 int	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
